+++ lib/active_interaction/active_job/core.rb
@@ -1,19 +1,11 @@
 module ActiveInteraction::ActiveJob::Core
   extend ActiveSupport::Concern
 
-  included do
-    define_job_class
-  end
-
   module ClassMethods
-    def define_job_class klass = default_job_class
+    def define_job_class klass
       unless const_defined?(:Job, false)
         const_set(:Job, Class.new(klass))
       end
-    end
-
-    def default_job_class
-      ActiveInteraction::ActiveJob::DefaultJob
     end
 
     def active_job &block
@@ -30,10 +22,8 @@ module ActiveInteraction::ActiveJob::Core
       subclass.define_job_class job_class
     end
 
-    def set options = {}
+    def delay options = {}
       ::ActiveInteraction::ActiveJob::ConfiguredJob.new(job_class, options)
     end
-
-    alias_method :delay, :set
   end
 end