+++ lib/active_interaction/active_job/sidekiq/default_job.rb
@@ -1,4 +1,0 @@
-class ActiveInteraction::ActiveJob::Sidekiq::DefaultJob
-  include Sidekiq::Worker
-  include ActiveInteraction::ActiveJob::Sidekiq::JobHelper
-end